# storage/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Desafio Api",
        "contact": {},
        "version": "1.0"
    },
    "servers": [{
        "url": "http://desafio-api.teste/api",
        "variables": {}
    }],
    "paths": {
...
        "/users/20000": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Destroy",
                "operationId": "Destroy",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "headers": {}
                    }
                },
                "deprecated": false,
                "security": [{
                    "bearer": []
                }]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [],
...
}